Use only one job for local unit imports
